refactor(perl-module): split rename planner into srp modules (#0000)
